moving main method to be in chronological order, changing example text to lorem ipsum.
