+++ ClientSideKeyVaultKeyToCustomerManagedKey/exampleDataCreator/blobExample.txt
