+++ ClientSideKeyVaultKeyToCustomerManagedKey/exampleDataCreator/config.py
