+++ ClientSideKeyVaultKeyToCustomerProvidedKey/exampleDataCreator/blobExample.txt
