+++ ClientSideKeyVaultKeyToCustomerProvidedKey/exampleDataCreator/config.py
