+++ ClientSideKeyVaultKeyToCustomerProvidedKey/migration.py
@@ -1,10 +1,26 @@
 import base64
-from ClientSideKeyVaultKeyToCustomerProvidedKey.setup import config as cfg
+from ClientSideKeyVaultKeyToCustomerProvidedKey.exampleDataCreator import config as cfg
 from azure.keyvault.keys.crypto import CryptographyClient, KeyWrapAlgorithm
 from azure.storage.blob import BlobServiceClient
 from azure.keyvault.keys import KeyClient, KeyVaultKey, KeyType
 from azure.identity import ClientSecretCredential
 from azure.keyvault.secrets import SecretClient
+
+
+def main():
+    # credential required to access client account
+    credentials = ClientSecretCredential(cfg.TENANT_ID, cfg.CLIENT_ID,
+                                         cfg.CLIENT_SECRET)
+    # access blob client with connection string
+    bs_client = BlobServiceClient.from_connection_string(cfg.connection_str)
+    key_client = KeyClient(vault_url=cfg.KEYVAULT_URL, credential=credentials)
+    # access container by name-- required that container already exists
+    cont_client = bs_client.get_container_client(cfg.cont_name)
+
+    # call to methods
+    key_vault_key = get_keyvault_key(credentials, key_client)
+    download_blob(cfg.blob_name, cont_client, key_vault_key, credentials)
+    upload_blob(bs_client, cfg.cont_name, cfg.blob_name)
 
 
 class KeyWrapper:
@@ -78,22 +94,6 @@ def upload_blob(blob_service_client, cont_name, blob_name):
     # upload contents to that blob with customer provided key for server side encryption
     blob_client.upload_blob(file_content, cpk=cfg.customer_key,
                             blob_type=b_type, overwrite=True)
-
-
-def main():
-    # credential required to access client account
-    credentials = ClientSecretCredential(cfg.TENANT_ID, cfg.CLIENT_ID,
-                                         cfg.CLIENT_SECRET)
-    # access blob client with connection string
-    bs_client = BlobServiceClient.from_connection_string(cfg.connection_str)
-    key_client = KeyClient(vault_url=cfg.KEYVAULT_URL, credential=credentials)
-    # access container by name-- required that container already exists
-    cont_client = bs_client.get_container_client(cfg.cont_name)
-
-    # call to methods
-    key_vault_key = get_keyvault_key(credentials, key_client)
-    download_blob(cfg.blob_name, cont_client, key_vault_key, credentials)
-    upload_blob(bs_client, cfg.cont_name, cfg.blob_name)
 
 
 if __name__ == "__main__":
