+++ ClientSideKeyVaultKeyToMicrosoftManagedKey/exampleDataCreator/blobExample.txt
