+++ ClientSideLocalKeyToCustomerManagedKey/exampleDataCreator/blobExample.txt
