+++ ClientSideLocalKeyToCustomerProvidedKey/exampleDataCreator/blobExample.txt
