+++ ClientSideLocalKeyToMicrosoftManagedKey/exampleDataCreator/blobExample.txt
