+++ ClientSideLocalKeyToMicrosoftManagedKey/exampleDataCreator/exampleDataCreator.py
@@ -1,8 +1,26 @@
 import os
 from azure.storage.blob import BlobServiceClient
 from azure.keyvault.keys import KeyClient
-from ClientSideLocalKeyToCustomerProvidedKey.setup import config as cfg
+from ClientSideLocalKeyToMicrosoftManagedKey.exampleDataCreator import config as cfg
 from azure.identity import ClientSecretCredential
+
+
+def main():
+    credentials = ClientSecretCredential(cfg.TENANT_ID, cfg.CLIENT_ID, cfg.CLIENT_SECRET)
+    # access a container using connection string
+    bs_client = BlobServiceClient.from_connection_string(cfg.connection_str)
+    key_client = KeyClient(vault_url=cfg.KEYVAULT_URL, credential=credentials)
+
+    # create container by name
+    try:
+        cont_client = bs_client.create_container(cfg.cont_name)
+    except:
+        cont_client = bs_client.get_container_client(cfg.cont_name)
+
+    # call to methods
+    create_encryption_scope()
+    content = get_content(cfg.blob_name)
+    upload_blob(content, bs_client, cfg.cont_name, cfg.blob_name)
 
 
 class KeyWrapper:
@@ -30,6 +48,12 @@ class KeyWrapper:
         return self.kid
 
 
+def create_encryption_scope():
+    print("\nCreating Microsoft Managed Key Encryption Scope...\n")
+    os.system(
+        'cmd /c "az storage account encryption-scope create --account-name ' + cfg.STORAGE_ACCOUNT + ' --name ' + cfg.serverside_managed_encryption_scope + ' --key-source Microsoft.KeyVault --resource-group ' + cfg.RESOURCE_GROUP + ' --subscription ' + cfg.SUB_ID + '"')
+
+
 def get_content(filename):
     file_content = open(filename, "rb+")
     data = file_content.read()
@@ -49,23 +73,6 @@ def upload_blob(data, blob_service_client, container_name, b_name):
     print("\nEncrypting blob on server...")
 
     blob_client.upload_blob(data, overwrite=True)
-
-
-def main():
-    credentials = ClientSecretCredential(cfg.TENANT_ID, cfg.CLIENT_ID, cfg.CLIENT_SECRET)
-    # access a container using connection string
-    bs_client = BlobServiceClient.from_connection_string(cfg.connection_str)
-    key_client = KeyClient(vault_url=cfg.KEYVAULT_URL, credential=credentials)
-
-    # create container by name
-    try:
-        cont_client = bs_client.create_container(cfg.cont_name)
-    except:
-        cont_client = bs_client.get_container_client(cfg.cont_name)
-
-    # call to methods
-    content = get_content(cfg.blob_name)
-    upload_blob(content, bs_client, cfg.cont_name, cfg.blob_name)
 
 
 if __name__ == "__main__":
